Added component for PDF creation
